requirements.txt
# =======================
# AI弹窗项目 - 依赖清单
# 最后更新: 2026-04-15
# 状态: 生产就绪 ✅
# Python版本要求: >=3.10.0 (请确保Python版本符合要求)
# =======================

# =======================
# GUI框架 - PyQt5
# =======================
PyQt5>=5.15.10
PyQt5-sip>=12.13.1

# =======================
# Web框架 - FastAPI
# =======================
fastapi>=0.120.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.6
pydantic>=2.10.0
pydantic-settings>=2.5.0

# =======================
# Web界面相关
# =======================
jinja2>=3.1.4
aiofiles>=24.0.0
websockets>=14.0
python-socketio>=5.12.0
schedule>=1.2.2

# =======================
# AI和机器学习核心
# =======================
# PyTorch (CPU版本 - 生产环境推荐)
torch>=2.5.0
torchvision>=0.20.0
torchaudio>=2.5.0

# 本地LLM支持
ollama>=0.3.0

# =======================
# 人脸识别和处理
# =======================
# InsightFace - 人脸分析核心 (固定版本保证稳定性)
insightface==0.7.3

# 人脸识别库
face-recognition>=1.3.0
dlib>=19.24.0

# 深度人脸分析
deepface>=0.0.90

# 面部网格检测
mediapipe>=0.10.15

# ONNX推理加速 (兼容性版本)
onnxruntime>=1.17.0

# 文本生成和扩散模型
transformers>=4.46.0
diffusers>=0.31.0

# =======================
# 图像和视频处理
# =======================
opencv-python>=4.10.0
opencv-python-headless>=4.10.0
Pillow>=11.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.14.0
imageio>=2.36.0

# 视频处理
moviepy>=2.0.0
imageio-ffmpeg>=0.5.0

# =======================
# 数据处理和科学计算
# =======================
pandas>=2.2.0
matplotlib>=3.9.0
scikit-learn>=1.5.0
scikit-image>=0.24.0

# =======================
# 异步和并发处理
# =======================
concurrent-log-handler>=0.9.25
threadpoolctl>=3.5.0

# =======================
# 网络和异步处理
# =======================
httpx>=0.28.0
aiohttp>=3.10.0

# =======================
# 数据序列化和配置
# =======================
orjson>=3.10.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
toml>=0.10.2

# =======================
# 文件和系统工具
# =======================
watchdog>=6.0.0
pathvalidate>=3.2.1
tqdm>=4.67.0
psutil>=6.1.0

# =======================
# 日志系统
# =======================
loguru>=0.7.3
colorlog>=6.9.0

# =======================
# HTTP客户端
# =======================
requests>=2.32.0
urllib3>=2.3.0

# =======================
# 安全相关
# =======================
cryptography>=43.0.0
bcrypt>=4.2.0
passlib>=1.7.4

# =======================
# 测试框架
# =======================
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
pytest-mock>=3.14.0

# =======================
# 代码质量工具
# =======================
black>=25.0.0
isort>=6.0.0
flake8>=7.1.0
mypy>=1.12.0

# =======================
# 实用工具
# =======================
click>=8.1.7
rich>=13.9.0
typer>=0.15.0

# =======================
# 平台特定依赖
# =======================
# Windows特定依赖
pywin32>=308; sys_platform == 'win32'

# macOS特定依赖
pyobjc>=10.0; sys_platform == 'darwin'

# =======================
# GPU支持 (可选 - 根据需要取消注释)
# =======================
# CUDA 12.1 支持 (生产环境请根据硬件选择)
# torch>=2.5.0+cu121; sys_platform == 'linux' or sys_platform == 'win32'
# torchvision>=0.20.0+cu121; sys_platform == 'linux' or sys_platform == 'win32'
# torchaudio>=2.5.0+cu121; sys_platform == 'linux' or sys_platform == 'win32'
# onnxruntime-gpu>=1.20.0; sys_platform == 'linux' or sys_platform == 'win32'

# DirectML支持 (Windows)
# torch-directml>=0.2.5; sys_platform == 'win32'

# =======================
# 开发和调试工具 (仅开发环境)
# =======================
pip-tools>=7.4.0
pre-commit>=4.0.0
ipykernel>=6.30.0
